requirements.txt
RPi.GPIO==0.7.0
python-dotenv==1.0.0
mfrc522==0.0.7
spidev==3.6
python-dateutil==2.8.2